Fixed a lot of issues in multiple files and working on main
The issues where mostly NULL references, which are illegal in C++, missing references to classes that messages have to be send to and the calling of non-static functions using the syntax: classType.function(). These have been replaced with references to the classes, so now: referenceToClassInstance.function().

First commit if main and I have, theoretically, fixed a circular reference encountered in the classes UpdateGametimeControl and ProcessHitControl.